refactor(telephony): make vobiz provider self-contained
Move the Vobiz REST client (MakeCall) from pkg/clients/vobiz into
telephony/internal/vobiz/internal so the provider is self-contained like
twilio/telnyx — no shared pkg dependency, uses net/http directly. Keeps
all provider code within telephony/internal/{provider} per review.
