+++ api/assistant-api/internal/channel/telephony/internal/vobiz/internal/client_test.go
@@ -3,7 +3,7 @@
 //
 // Licensed under GPL-2.0 with Rapida Additional Terms.
 // See LICENSE.md or contact sales@rapida.ai for commercial usage.
-package vobiz
+package internal_vobiz
 
 import (
 	"context"
